Add `parallel` option for mviews
Adding a custom configuration option for Oracle materialized views,
`parallel`. When set like this

```
create_view(:mview_name, materialized: {parallel: true})
```

Then the mview will be created with

`CREATE MATERIALIZED VIEW mview_name PARALLEL...`

What is `parallel`?

From Oracle Docs

> The parallel_clause lets you parallelize creation of the table and set the default degree of parallelism for queries and the DML INSERT, UPDATE, DELETE, and MERGE after table creation.
...
> Specify PARALLEL if you want Oracle to select a degree of parallelism equal to the number of CPUs available on all participating instances times the value of the PARALLEL_THREADS_PER_CPU initialization parameter.

In my testing this setting has led to faster mview refresh times. But I
wasn't sure about the impact of using it. So I asked Andy Wattenhofer,
an expert on Oracle. (emphasis mine)

---

Me: Is there a reason I wouldn't want to always use it?

Andy: It's not a simple answer, but **you would probably always want to use the parallel clause for large datasets**. It could cause problems if everyone were to use it at the same time. You're effectively creating several server threads to perform the data load in parallel chunks. It would be possible to saturate one or more of the database servers if it were used too heavily. For that reason I would suggest that you try specifying an integer with it, such as parallel 8 or parallel 16 and fine tune it from there to find the value that maximizes on time for the smallest number of threads...I would request that you be careful in your use of too much parallelism.

---

In this version I'm not providing the integer paramater (i.e.,
`parallel` 8), instead just using `parallel`. This will always use the
max number of CPUS. See the above Oracle doc excerpt.

The suggestion is that this should only be used on large, slow
materialized views.