Refine `view_dependencies` query to return user-controlled objects
Closes #18

As mentioned in https://github.com/cdinger/scenic-oracle_adapter/issues/18#issuecomment-2206657472
this updates the query so that it only returns a dependency if it exists
in `user_objects`. So dependencies such as `user_scheduler_jobs` or
other Oracle-controlled objects are excluded.

Also fixing a weird edge case. If a view/mview has a dependency on an object
in a different schema that happens to have the same name as a
user-created object in this schema, then we want to exclude it.