Provide fallback view sorting
This provides a default fallback sorting if a view is missing from the
increasingly complex `view_dependencies` query. We've already
encountered several edge cases were views are missing or misclassified
in Oracle's `user_dependencies` view.

If a new edge case pops up this should not explode on a failed index
lookup. Instead it will be added to the end of the sorted views array.