Check control file when detecting available DB
I'm hoping this makes the tests via Github Actions more reliable.
They've been randomly failing because the test suite starts executing as
soon as the database allows connections.

A better test seems to be querying the database to ensure it's ready for
queries/etc. This is done by querying metadata about the control file
(according to this stack overflow question):

https://serverfault.com/questions/38488/how-can-i-tell-if-an-oracle-database-is-mounted-and-activated