Generate a consistent schema.rb file
**Problem**

When loading an existing schema that contains `create_view` statements
and then dumping the schema (e.g., if a test script calls `db:migrate`
before running tests), the `schema.rb` file changes even though the
database objects remain the same. This inconsistency can also occur when
tests use `ActiveRecord::Migration.maintain_test_schema!`. This issue
may be related to or resolve #23.

**Example**

Suppose we start with a `schema.rb` that contains a Scenic view:
```
  create_view "my_view", sql_defintion: <<-SQL
      SELECT
        1 AS value
      FROM
        dual
  SQL
```

The `sql_definition`, is a multi-line string:
`"        SELECT\n        1 AS value\n      FROM\n        dual\n"`

When `schema:load` is executed, it creates the view using
`trimmed_definition` which removes leading whitespace but preserves the
indentation of subsequent lines. As a result, the view in Oracle is:
```
SELECT
        1 AS value
      FROM
        dual
```

After dumping the schema (e.g., after a migration), the view appears as:
```
  create_view "my_view", sql_definition: <<-SQL
      SELECT
          1 AS value
        FROM
          dual
  SQL
```
Notably:
  * The first line (`SELECT`) matches the original formatting.
  * Subsequent lines are indented two more spaces than before.

**Cause**

The issue arises from Scenic's `to_schema` method, which uses a heredoc
as a template for the `create_view` statement and applies `indent(2)` to
the view definition:
```
    def to_schema
      materialized_option = materialized ? "materialized: true, " : ""

      <<-DEFINITION
  create_view #{UnaffixedName.for(name).inspect}, #{materialized_option}sql_definition: <<-\SQL
    #{escaped_definition.indent(2)}
  SQL
      DEFINITION
    end
```

The `indent(2)` adds two leading spaces to each line of the view
definition. While `trimmed_definition` removes leading spaces from the
first line, it does not adjust the indentation of subsequent lines. This
results in extra spaces accumulating in the `schema.rb` file each time
the schema is loaded and dumped.

**Solution**

To fix this we:
  1. Updated `trimmed_definition`: We replaced `strip` with
     [`strip_heredoc`](https://api.rubyonrails.org/classes/String.html#method-i-strip_heredoc) and `rstrip` to remove a consistent amount of
     leading whitespace from all lines (if any) and trailing whitespace
     from the definition.
  2. Added a custom `View` Class: We introduced a `View` class to the
     Scenic Oracle adapter with a custom `to_schema` method. This
     addresses formatting issues caused by inconsistent indentation
     between the first line and the rest due to the `heredoc` and
     `indent(2)`.

**Testing**

We added a test to simulate this behavior. One implementation detail for
this test is a helper method in `spec_helper` to dump the schema. This
ensures we get a fresh dumper each time because the dumper seems to
cache the database state from when it was initialized, so reusing it
does not detect schema changes.

Co-authored-by: Davin Lagerroos <dlagerro@umn.edu>
Co-authored-by: Eric Eklund <eeklund@umn.edu>