Require tsort before scenic
The Scenic::Adapters::Postgres::Views::TSortableHash class is including
TSort, but doesn't have a require statement. When we `require "scenic"`,
it fails with:

```
uninitialized constant
Scenic::Adapters::Postgres::Views::TSortableHash::TSort
```

Adding this before scenic ensure TSort is available. We should open a PR
to scenic to add this there.