Include no_data mview information in schema.rb
This is not implemented in Scenic itself. Postgres doesn't have a way to
know if a materialized view was created `WITH NO DATA`; if can only
detect if a view is populated. Oracle however does store the
`BUILD_MODE` on mviews created with `BUILD DEFERRED`.

This adds a `Scenic::Adapters::Oracle::View` class that extends
`Scenic::View` with this information so it can be dump/loaded via
schema.rb